backend/requirements.txt
ghostscript
selenium
flask
argparse
bs4
gunicorn
flask_cors
flask_apscheduler
pydrive2
python-dotenv
lxml
behave
assertpy
celery